bumping version to 1.3.12, and fixing some gcc x64 bugs
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@1034 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ test_suite/test_stde_extended_cstring.cpp
@@ -69,8 +69,8 @@ void testStdeSafeFormat()
 	LASS_TEST_CHECK_EQUAL(safe_format("%s %s!", "hello", "world"), "hello world!");
 
 	const char* a = "12345678901234567890123456789012345678901234567890";
-	LASS_TEST_CHECK_EQUAL(strlen(a), 50);
-	LASS_TEST_CHECK_EQUAL(safe_format("%s%s%s%s%s%s%s%s%s%s", a, a, a, a, a, a, a, a, a, a).length(), 500);
+	LASS_TEST_CHECK_EQUAL(strlen(a), size_t(50));
+	LASS_TEST_CHECK_EQUAL(safe_format("%s%s%s%s%s%s%s%s%s%s", a, a, a, a, a, a, a, a, a, a).length(), size_t(500));
 }
 
 TUnitTest test_stde_extended_cstring()
